[refactor] polymarket and dashboard components into modular components
