+++ src/sections/banking/polymarket-activity-context.tsx
@@ -361,4 +361,3 @@ export function usePolymarketActivity(): ActivityContextValue {
   }
   return ctx
 }
-
